Moved scripts into utils
